copy consensus files
